Create new useSuggestedFollowsByActorWithDismiss hook
